requirements.txt
Flask
psycopg[binary]
psycopg-pool
python-dotenv
flask-socketio
cloudinary
